terminado lista-platos corroborar que tengo todo y limpiar codigo

+++ src/helper/menuProvider/informationMenu.js
@@ -7,7 +7,7 @@ function informationInteres(platos, firstAttribute, secondAttribute = null) {
   }, 0);
 
   return secondAttribute
-    ? parseInt(preparationMinutes.toFixed(2))
+    ? parseFloat(preparationMinutes.toFixed(2))
     : parseInt(preparationMinutes);
 }
